requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0
streamlit>=1.30.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
scikit-learn>=1.3.0
albumentations>=1.3.0
opencv-python-headless>=4.8.0
tqdm>=4.66.0
tensorboard>=2.15.0
pyyaml>=6.0